pyproject.toml
[build-system]
requires = ["python>=3.7", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "URL Summarizer"
version = "0.1.0"
description = "Your Flask app description"
authors = ["Gautam Pahwa"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.0.0"
newspaper3k = "^0.2.8"
werkzeug = "^2.0.0"

[tool.poetry.dev-dependencies]

